pyproject.toml
[project]
name = "beimeng-workspace"
version = "0.1.0"
description = "AI-Friendly Python Monorepo - 包含多种应用、脚本和文档的工作空间"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Beimeng Team" }
]

# 核心依赖
dependencies = [
    "typer>=0.12.0", # CLI 框架
    "pydantic>=2.0.0", # 数据验证
    "pydantic-settings>=2.0.0", # 配置管理
    "rich>=13.0.0", # 终端美化
    "loguru>=0.7.0", # 日志管理
    "python-dotenv>=1.0.0", # 环境变量
    "pyyaml>=6.0", # YAML 支持
    "uvicorn>=0.38.0",
    "fastapi>=0.123.0",
    "python-multipart>=0.0.20",
    "itsdangerous>=2.2.0",
    "httpx>=0.28.1",
    "playwright>=1.49.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
    "jinja2>=3.1.6",
    "chardet>=5.2.0",
    "playwright-stealth>=2.0.0",
]

[project.optional-dependencies]
# 开发工具
dev = [
    "ruff>=0.3.0",             # Linting + Formatting
    "mypy>=1.9.0",             # 类型检查
    "pytest>=8.0.0",           # 测试框架
    "pytest-cov>=4.1.0",       # 测试覆盖率
    "pytest-asyncio>=0.23.0",  # 异步测试
    "pre-commit>=3.6.0",       # Git hooks
    "pyinstaller>=6.11.0",     # 打包工具
    "psutil>=5.9.0",           # 系统监控（测试依赖）
    "aiohttp>=3.9.0",          # 异步 HTTP（测试依赖）
]

# 文档工具
docs = [
    "mkdocs>=1.5.0",           # 文档生成
    "mkdocs-material>=9.5.0",  # Material 主题
    "pdoc>=14.0.0",            # API 文档
]

# Web 应用
web = [
    "fastapi>=0.110.0",        # Web 框架
    "uvicorn>=0.27.0",         # ASGI 服务器
    "httpx>=0.27.0",           # HTTP 客户端
    "jinja2>=3.1.0",           # 模板渲染
    "python-multipart>=0.0.9", # 表单上传
]

# 数据处理
data = [
    "pandas>=2.2.0",           # 数据分析
    "polars>=0.20.0",          # 高性能数据框架
]

# Temu 自动发布
temu = [
    "pandas>=2.2.0",           # Excel 处理
    "openpyxl>=3.1.0",         # Excel 读写
    "playwright>=1.56.0",      # 浏览器自动化
    "playwright-stealth>=1.0.0",  # 反检测
    "tenacity>=8.2.3",         # 稳定重试
    "syncer>=2.0.0",           # 将异步任务同步执行
    "selenium>=4.21.0",        # 备用 Selenium 认领流程
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Hatchling 配置 - Monorepo 打包
[tool.hatch.build.targets.wheel]
packages = ["apps", "scripts", "packages"]

# Ruff 配置 - 现代化的 Python linter 和 formatter
[tool.ruff]
line-length = 100
target-version = "py312"

# 选择要启用的规则
[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # 允许在参数默认值中调用函数 (Typer 需要)
    "N999",   # 允许非标准模块名 (airtest 等)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # 允许未使用的导入
"tests/*.py" = ["RUF059", "E402"]  # 测试中允许未使用的变量和非顶部导入
"**/cli/commands/*.py" = ["E402"]  # CLI命令允许非顶部导入(需要先加载环境变量)
"**/service.py" = ["RUF059"]  # service 文件允许未使用的变量
"**/test_*.py" = ["B007"]  # 测试文件允许循环变量未使用
"**/examples/*.py" = ["E402"]  # 示例文件允许非顶部导入

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# MyPy 配置 - 类型检查
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

# Pytest 配置
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=apps/temu-auto-publish/web_panel",
    "--cov=apps/temu-auto-publish/config",
    "--cov=packages/common",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
pythonpath = ["."]

# Coverage 配置
[tool.coverage.run]
source = [
    "apps/temu-auto-publish/web_panel",
    "apps/temu-auto-publish/config",
    "packages/common",
]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Bandit 配置 - 安全扫描
[tool.bandit]
exclude_dirs = ["tests", "build", "dist", ".venv", "venv"]
skips = ["B101"]  # 排除 assert 检查（测试中常用）

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.7",
    "psutil>=5.9.0",
    "aiohttp>=3.9.0",
]
